pyproject.toml
[project]
name = "extract-gmail-attachment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fsspec>=2025.3.0",
    "google-api-python-client>=2.163.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "pytest>=8.3.5",
]
